amqp_client: add connect_timeout
Add a connect timeout, and socket_timeout for pika. This means each
connection attempt will take a limited amount of time, and after
connect_timeout time has passed, we'll throw an error.

In case of a blocking .consume(), this error will simply be raised.
In case of .consume_in_thread, the error must be extracted out to the
calling thread, and raised there.