Add resuming to LifecycleProcessor graphs
Refactor the LifecycleProcessor a bit so that the graph is actually
returned from a function, so that we can use the decorator. Only
private method behaviour can be changed here, because plugins do
rely on it.

Also add a name_prefix so that each graph in a workflow is stored
with a unique name, because some workflows do call .install()
multiple times (eg. the update workflow)