CY-844 Add resuming to builtin workflows  (#176)
* CY-844 Add a decorator for automatically handling store/restore

Workflows can now use this decorator around a function which
creates the tasks graph so that it is automatically stored and
restored when needed

* Add stubs for local (local is not supported for restoring yet)

* CY-844 Add resuming to execute_operation

* Add resuming to install_new_agents

* Add resuming to LifecycleProcessor graphs

Refactor the LifecycleProcessor a bit so that the graph is actually
returned from a function, so that we can use the decorator. Only
private method behaviour can be changed here, because plugins do
rely on it.

Also add a name_prefix so that each graph in a workflow is stored
with a unique name, because some workflows do call .install()
multiple times (eg. the update workflow)

* For consistency, just graph here

All the other calls next to it do the same. Anyway, ctx.graph_mode()
currently always returns the same object anyway

* TasksGraph: don't execute tasks that are already pending

* MethodDecorator unittest: add args

* make_or_get_graph: builtin compat with methods

No need for @method_decorator

* install_new_agents: pass name by kwarg