CY-1083
I am not sure if this is the right AMI. The old AMI was ami-c1c699a7. Now it is no longer available. Currently this region offers me an ARM and an x86 version. I'm guessing we need the x86 version since that's what we apparently used according to the old wagon names, but I'm not sure. I'm also not sure if this API is available in the account Jenkins uses. I guess we will need to change this and run the build again and see if it gets fixed.