amqp_client: call handlers' register in-thread
When client.add_handler is called after the connection has
already been made, the register function was called in the
calling thread, not in the connection thread.

This made it so that we were avoiding that mode of .add_handler
(always calling it before connecting instead), because the alternative
was unsecured multithreaded access to the connection.

With this change, register is always called from the connection
thread when .add_handler is used, which makes it safe to add handlers
after the connection has already been made.