RequestResponseHandler: use the correlation id as response queue name
The response queue was mostly an artificial thing, and we already
have a unique identifier - the correlation id. We can use that, and
there's one thing less to keep track of.

This is allowed now because the queues are created on-demand when
a task is sent (when the corr. id is known), not up-front.

History: the original idea to have a separate response queue name
was that maybe multiple task responses could be received there.
However it wasn't really feasible in face of multiple consumers from
that queue. Also, after a resume we want to connect back to a queue
for a specific task, not for other tasks. And so, a known-up-front
queue name wasn't really ever able to be taken advantage of.