Only Blocking handler sets exclusive=true
Blocking calls can't be resumed after a process restart, so there's
no point keeping the queue around - nobody is going to read it.

Therefore, set exclusive=true in it, and the queue will be deleted
automatically by rabbitmq if the consumer process crashes
(when the connection closes).

CallbackHandler calls can be resumed and so for that, we keep the
default exclusive=false, so that the queue isn't deleted