BlockingRequestResponseHandler call: arg name
We removed __init__ from _RequestResponseHandlerBase, so the param
names are now from TaskConsumer, and for that, the arg name is
`queue`. Even though it's really an exchange name, not a queue.

I decided that this is bullshit, and the safe thing to do is to just
decide that the exchange name shall be passed positionally, not by
name.
And in the future we'll also rename the arg.

This means all uses of _RequestResponseHandlerBase's subclasses
must be audited, and at least a cloudify-agent PR must accompany this.

History: we _did_ know the difference between a rabbitmq exchange
and a queue when we wrote this, but celery mostly didn't (because
celery is somewhat-broker-agnostic and the primary user-facing thing
in celery is a "queue"). And when we were implementing this, we
inherited most of celery's naming, because of all the envvars etc
we had around.