RequestResponseHandlers: make sure amqp queues are deleted (#320)
* Declare queues on demand

Instead of creating them up-front, only create the queue when we
actually send a request

* RequestResponseHandler: use the correlation id as response queue name

The response queue was mostly an artificial thing, and we already
have a unique identifier - the correlation id. We can use that, and
there's one thing less to keep track of.

This is allowed now because the queues are created on-demand when
a task is sent (when the corr. id is known), not up-front.

History: the original idea to have a separate response queue name
was that maybe multiple task responses could be received there.
However it wasn't really feasible in face of multiple consumers from
that queue. Also, after a resume we want to connect back to a queue
for a specific task, not for other tasks. And so, a known-up-front
queue name wasn't really ever able to be taken advantage of.

* remote apply_async: use is_agent_alive

Now that we have `connect=False` for is_agent_alive, and threadsafe
.register, we can use the client we passed there safely.

This means we can use the nice public api is_agent_alive instead
of constructing a handler manually.

* Only Blocking handler sets exclusive=true

Blocking calls can't be resumed after a process restart, so there's
no point keeping the queue around - nobody is going to read it.

Therefore, set exclusive=true in it, and the queue will be deleted
automatically by rabbitmq if the consumer process crashes
(when the connection closes).

CallbackHandler calls can be resumed and so for that, we keep the
default exclusive=false, so that the queue isn't deleted

* Simplify BlockingRequestResponseHandler

It's blocking, so it can only ever work with a single req/resp
at a time, so there's no point keeping a dict of queues.
Just keep a single queue.

* Reorganize CallbackRequestResponseHandler.process

Just a minor refactor and moving things around. Saves the loads if
there's nobody listening. It'll make the next changes easier.

* RequestResponseHandlers: delete the queue when a message is received

Response queues will only ever receive one message, so when we
get it, we can delete the queue right away.

This prevents queue leaks in some complex scenario that i don't really
remember well (the test scenario for the 4.6-resume-cancelled patch)

* wait=False when deleting queues

.process is on the connection thread, so we can't wait, that'd just
deadlock

* BlockingRequestResponseHandler call: arg name

We removed __init__ from _RequestResponseHandlerBase, so the param
names are now from TaskConsumer, and for that, the arg name is
`queue`. Even though it's really an exchange name, not a queue.

I decided that this is bullshit, and the safe thing to do is to just
decide that the exchange name shall be passed positionally, not by
name.
And in the future we'll also rename the arg.

This means all uses of _RequestResponseHandlerBase's subclasses
must be audited, and at least a cloudify-agent PR must accompany this.

History: we _did_ know the difference between a rabbitmq exchange
and a queue when we wrote this, but celery mostly didn't (because
celery is somewhat-broker-agnostic and the primary user-facing thing
in celery is a "queue"). And when we were implementing this, we
inherited most of celery's naming, because of all the envvars etc
we had around.

* Also delete the queues if they're nonempty

It doesn't make sense for there to be additional messages in
the queues. Nobody is going to read them. Just delete the queue.