CY-1043 Get rabbitmq credentials from rest when sending tasks
Instead of storing the credentials in operation context, which
might be stored in places that we don't want to store credentials in
(eg. the operations table), get the credentials via rest when
sending.

Also declare the credential fields on the restclient Tenant object -
they were there before (since CY-599), but undeclared