Figure out the node at runtime
GetProperty needs a way to figure out the node both at runtime and
at parse time. The context at runtime is slightly different than
at parse time, so we need to handle both cases.

It's ugly that the cases are different and need to be handled
differently. Making them more similar/the same, is outside the scope
of this issue (it is tracked in CY-1676 instead)