Change the nested-circular function call validation
Now the validation is done in prepare_deployment_plan instead of
in parse.
This means functions do it in evaluate - in parse time they expect
all the dependency functions to already be resolved, so if there's
some function that wasn't evaluated in the arguments list, that's an
error.

Also update the tests to split out parse & prepare_deployment_plan
so that it's better visible where the assert is