CY-1507 Allow get_input and get_property at runtime (#356)
* Add the runtime_only_evaluation parameter for deployments

Add the parameter and pass it to the PlanHandler

* Simple refactors - flake etc

not important

* Evaluate functions at runtime

If the flag is set, PlanHandler evaluates functions at runtime.

This is the main part! The rest is just support :)

* Figure out the node at runtime

GetProperty needs a way to figure out the node both at runtime and
at parse time. The context at runtime is slightly different than
at parse time, so we need to handle both cases.

It's ugly that the cases are different and need to be handled
differently. Making them more similar/the same, is outside the scope
of this issue (it is tracked in CY-1676 instead)

* Change the nested-circular function call validation

Now the validation is done in prepare_deployment_plan instead of
in parse.
This means functions do it in evaluate - in parse time they expect
all the dependency functions to already be resolved, so if there's
some function that wasn't evaluated in the arguments list, that's an
error.

Also update the tests to split out parse & prepare_deployment_plan
so that it's better visible where the assert is

* Add evaluate_node_functions and evaluate_node_instance_functions

Pretty much utility functions, but part of the api of the module.
This is what the cloudify-manager side will call

* Add local workflows support for evaluating node and node instances

This was already supported by the manager

* Add a mock

* Style changes

Addressing comments

* Add additional unittests

* move import