apply lib2to3.fixes.fix_basestring
This rewrites `basestring` with just `str`.
Instead of doing just that, we declare a `text_type` in _compat, and
use that.
text_type is supposed to be the actual text type, ie. str in py3 and
unicode in py2. However, for now, in py2, it is both str and unicode,
because of the thousand tests who all pass bytes around.
This is not the final change, and we will drop `str` from there
in the future.