rewrite the schema type validator check
Now that we have a eg. `schema = (int, text_types)`, where
text_types is a tuple itself (on py2) the primitive schema type
validation doesn't work anymore, because it expects no nesting
of tuples.

However, isinstance() is just fine with that, so there's no reason
the validation check to be that strict.
And there's no reason to have a separate validation check anyway,
if this is to be used with isinstance, let's just check the same
way.