CY-2367 Py3 porting: step 7 (#417)
* apply lib2to3.fixes.fix_basestring

This rewrites `basestring` with just `str`.
Instead of doing just that, we declare a `text_type` in _compat, and
use that.
text_type is supposed to be the actual text type, ie. str in py3 and
unicode in py2. However, for now, in py2, it is both str and unicode,
because of the thousand tests who all pass bytes around.
This is not the final change, and we will drop `str` from there
in the future.

* Make the dsl-parser load unicode

dsl_parser used to force the strings to be bytes, but that
Just Wont Fly with py3, and really, there is no reason to do so.
Revert back to the original behaviour of yaml (and json), which is
to yield unicode strings

* rewrite the schema type validator check

Now that we have a eg. `schema = (int, text_types)`, where
text_types is a tuple itself (on py2) the primitive schema type
validation doesn't work anymore, because it expects no nesting
of tuples.

However, isinstance() is just fine with that, so there's no reason
the validation check to be that strict.
And there's no reason to have a separate validation check anyway,
if this is to be used with isinstance, let's just check the same
way.

* rewrite error messages in tests

All of those super-tightly-coupled error messages are making my
head hurt. Checking that the message contains `<type "str">`?!

* fix script_runner tests

this only really relates to the tests, who decide whether or not
use json based on if the input value is a string

* fix local workflows tests

again, with the error message checking. This is nicer.

* local workflows type checking

local workflows have their own typechecker for inputs, but all the
tests do provide bytes, so lets (for now, similar to the change in
dsl-parser) allow bytes as well.

Additionally, no need to convert from-string-to-string

* add fix_basestring to fixers list