drop test_download_resource_to_non_writable_location
whether or not that location is writable depends on whether the
tests are run as root or not. As root, it will happily create a file
named `/non-existing-folder`.

All in all, it doesn't really make sense to test that a file that
is not writable, is really not writable.  Unless we want to test
that we didn't find a magical way of bypassing unix permissions.
Which we don't.