tasks_graph: return a list for _terminated_tasks
the graph is mutated as we traverse the terminated tasks,
so we need to gather them up-front.

That happens only for terminated tasks, because only for them
the tasks are removed from the graph.

That list isn't going to be very big, because it only contains
the tasks that terminated in the current iteration.