CY-2411 Py3 porting: test fixes (#425)
* Replace exception.message with str(exception)

exception.message is deprecated in py27 and dropped in py3
Replace it with str(e).

In the one place where we do actually really need a message attribute,
make one explicitly.

* Use assertRaises/assertRaisesRegex

- _assert_dsl_parsing_exception_error_code now has a mode of checking
  message against a regex
- replace assertRaisesRegexp with assertRaisesRegex (it's a testtools
  method, and the "Regexp" naming is deprecated)

* Fix dsl-parser tests depending on ordering

Several tests do things like `nodes[0], nodes[1]` but nodes is
actually coming from a dict, and we can't guarantee the ordering
of it.
Instead of relying on indexing like that, where needed,
find the node using other means.

* dsl_parser: drop test_unicode_input

All inputs are unicode. It doesn't make sense to test it as a
separate case.
Inputs are not bytestrings, because they come from json/yaml

* test_validate_secrets_some_invalid: don't rely on ordering

The test used to rely on ordering of get_secret calls, and
hardcoded responses in order. Now, it'll just respond based on
the secret name, as required by the test

* proxy tests: drop test_ctx_in_path

Whether or not ctx is in path, depends how the package was installed
for testing.
For example, if you install via `pip install --user`, then it will
be put into ~/.local/bin, which isn't necessarily in the path.

* drop test_download_resource_to_non_writable_location

whether or not that location is writable depends on whether the
tests are run as root or not. As root, it will happily create a file
named `/non-existing-folder`.

All in all, it doesn't really make sense to test that a file that
is not writable, is really not writable.  Unless we want to test
that we didn't find a magical way of bypassing unix permissions.
Which we don't.

* event: verbosity level must be an int

in py3, it's invalid to compare `None > int`. It doesn't really
make sense to allow null values there, anyway

* Assert that deprecation warnings are thrown

.capabilities is deprecated (by us), so let's assert that a
deprecation warning is actually shown.

Otherwise, the deprecation warning bubbles up and becomes a
test warning

* tasks_graph: return a list for _terminated_tasks

the graph is mutated as we traverse the terminated tasks,
so we need to gather them up-front.

That happens only for terminated tasks, because only for them
the tasks are removed from the graph.

That list isn't going to be very big, because it only contains
the tasks that terminated in the current iteration.

* test_task_failed: fix the mock

it needs to actually have an int attribute, because otherwise
accessing .execute_after will return another Mock, and that cannot
be compared to an int in py3.

(another alternative would be to use MagicMock instead, but we
can be explicit here)

* test_local_workflows: make the mock module using types.Module

the imp module is deprecated. Instead, users are supposed to use
the importlib module, but that is incompatible with py2.

Fortunately, we only need a module object in sys.modules, with no
attributes, so we can just create one with types.Module

* flake

* skip the warnings asserts on py2.6

i have no idea how to make them work