Events: don't encode the message
Inside of events, the message is interpolated into a native string,
so there's no need to encode it.

Later, it is sent via rabbitmq, as json, ie. unicode, so encoding it
to bytes makes no sense at all.

Without this change, on py3, we see logs with the repr'd `b''` prefix