Pass objects as json in `ctx.returns()`
ctx.returns already used `-j` however the input data wasn't json,
but it is now.
This means that it is now actually possible to pass things back,
such as dicts, and they will not be stringified in a silly way. Also,
unicode objects can be passed, and ints, and...

For compatibility, things that cannot be jsonified (eg. like in the
tests, type/class objects), are repr'd and passed as string still.
That is silly, but it is backwards-compatible at least.

Since what is returned now is eg. a real dict, and not a string
representation of it, in ctx.returns we also pass an empty dict,
because otherwise the MutableMapping tango could want to use the
actual data as kwargs

In tests we can now assert actual objects and not ast.literal_eval
and other sillyness.
Two caveats with the test asserts:
  - it's <type int> on py2 but <class int> on py3, so just make the
    string comparison work with both
  - the test method we have returns `result or runtime_properties`,
    so falsy objects are wrapped in a {"value": ..} so that they
    can actually be asserted on (otherwise we'd get runtime props
    instead of the return value, in the test)

Note btw, `ctx.returns()` is a very underused feature because the
only way to really use it, is to have your custom workflow do
something like `value = operation.get()`, and there's no way of
using it with tasks graphs