CY-2430 Py3 porting: unicode/bytes changes (#426)
* Events: don't encode the message

Inside of events, the message is interpolated into a native string,
so there's no need to encode it.

Later, it is sent via rabbitmq, as json, ie. unicode, so encoding it
to bytes makes no sense at all.

Without this change, on py3, we see logs with the repr'd `b''` prefix

* get_resource/download_resource returns bytes

and so we must make sure we cope with bytes there.

In the future, we can think about whether it doesn't make more sense
to return unicode there instead

* subprocesses stdout/stderr return bytes

cope with bytes there, instead of unicode/native strings

* ctx proxy: pass bytes/unicode where needed

the zmq and http proxies have their assumptions where do bytes go
and where does unicode go, now python 3 forces us to actually
adhere to that

* use logger.warning instead of .warn

warn is deprecated

* Pass objects as json in `ctx.returns()`

ctx.returns already used `-j` however the input data wasn't json,
but it is now.
This means that it is now actually possible to pass things back,
such as dicts, and they will not be stringified in a silly way. Also,
unicode objects can be passed, and ints, and...

For compatibility, things that cannot be jsonified (eg. like in the
tests, type/class objects), are repr'd and passed as string still.
That is silly, but it is backwards-compatible at least.

Since what is returned now is eg. a real dict, and not a string
representation of it, in ctx.returns we also pass an empty dict,
because otherwise the MutableMapping tango could want to use the
actual data as kwargs

In tests we can now assert actual objects and not ast.literal_eval
and other sillyness.
Two caveats with the test asserts:
  - it's <type int> on py2 but <class int> on py3, so just make the
    string comparison work with both
  - the test method we have returns `result or runtime_properties`,
    so falsy objects are wrapped in a {"value": ..} so that they
    can actually be asserted on (otherwise we'd get runtime props
    instead of the return value, in the test)

Note btw, `ctx.returns()` is a very underused feature because the
only way to really use it, is to have your custom workflow do
something like `value = operation.get()`, and there's no way of
using it with tasks graphs

* py2+3-compatible error message

in py3 it's "FileNotFoundError", not "IOError", in the error message