CY-2900 start deployment modification: refresh node instances
In scale-like workflows (which can be custom workflows written
by the user!), it is always required to refresh node instances
after starting the scale, so that the local process gets the
newly-created node instances.

Without refreshing, it doesn't have them, and so it can't install
them.

Since it is always required, let's just make it happen automatically.