When checking for plugin dir, check for the marker file as well
Instead of just checking .isdir, also look for the plugin.id file.
Otherwise it will be just a directory and not a directory containing
a plugin.

In practice, it will be a plugin that is not fully installed yet,
so we should go to the installer, which will do the locking-and-waiting
dance.