Move plugin installer code to common (#535)
* Move plugin installer code to common

* Use common's plugin installer in dispatch

* Remove useless extract_package_name

This was only used for debug logs. I don't really want to subprocess
and do semi-unsupported stuff, in the name of debug logs.

* Drop _create_plugins_dir_if_missing

This isn't needed because
1. The dir is created in the rpm anyway
2. Creating a virtualenv will create it, in a `mkdir -p` fashion

Also note that it only created plugins/ and not source_plugins/,
and it didn't matter for source plugins. So we don't need to have
this for plugins/ either.

* Move PluginInstallationError here as well

* Rewrite PluginInstallationError's docstring

look how many useless words it had

* replace VIRTUALENV with sys.executable

VIRTUALENV is defined as dirname(sys.executable) so we might well
just use this directly

and with this change, there's no more cloudify_agent imports in this

* Inline _rmtree

Really no reason to have a wrapper

* Remove a few unused things

* Require wagon and fasteners