CY-3455 Add workflow_ctx.download_resource
Similar to ctx.download_resource.

This is because we sometimes _do_ need a download_resource on the
workflow level - specifically, when installing source plugins, we'll
use download_resource, on whichever context we're in - if it's an
operation function, then CloudifyContext, but if we're in a
workflow context, because we're installing a plugin containing the
workflow function, then we'll be using a WorkflowContext. So it
needs to be defined there as well.

But in principle any workflow function code can use download_resource.
And it's surprising this feature hasn't come up before.