CY-3534 manage_plugin_state fixes (#556)
* _install_managed_plugin: name the first argument `plugin`

This will make things a bit easier, because this allows us
to assume in the decorator that the plugin is just passed
positionally.

And it doesn't lose any information because we already know
it's a managed plugin, because we are in a install_managed
function.

* Wrap _install_managed_plugin in _manage_plugin_state

It's _this_ function that needs to update plugin state - the
other one works on both managed and source plugins, and it's
useless to try and set a source's plugin state. We only want
to do this for managed plugins (and it's only possible
for managed plugins, anyway, because only they have an id)

* get_daemon_name: remove unused arg

Just a whoops.