CY-3570 Get rid of the fileserver in unittests
Instead of starting a fileserver and using it in tests
for get_resource, mock out requests.get.

This makes the tests actual unittests, then.

Directly, this is because the port for the fileserver was hardcoded
and not changing, which meant we couldn't run multiple runs
of these tests at the same time, which led to random failures
on the new jenkins.