CY-3589 ClusterClient: copy the hosts list
We need to isolate ourselves from outside changes to the hosts list,
that's why we can copy it.

Main reason of the "outside changes" is another ClusterClient
running concurrently, and doing the shuffle() as well.

This can lead to the client contacting the same manager multiple
times, and skipping over some managers, depending on the shuffle luck.