get_resource: try all possible paths
get_resource is supposed to try 3 paths:
  - deployment-local  (for all managers)
  - blueprint-local  (for all managers)
  - manager-global  (for all managers)
..for up to `3 * len(managers)` calls in a cluster.

It used to, if one of the managers was offline, throw a NonRecoverable
on possibly the deployment-local stage.

This meant that if one manager was offline, it was impossible to
use get_resource to get blueprint-local or manager-global
resources.

Instead, make sure that get_resource never throws nonrecoverable,
always throws httpexception if it doesn't found, but most importantly-
always tries all managers, even if some managers are offline.
(I'll keep that `if` as 404 for now, I don't think that needs to
be made more lenient)