task .get: remove from graph
when we are NOT using tasks graphs, and we run a task, that task
is still added to a "tasks graph", even though we're not using one.

This is so that these tasks can be awaited for inside of dispatch.py,
_execute_workflow_function.

And so, when the task is done, remove it from that graph, so that
we don't need to await on it anymore.

This is unfortunate and probably stupid, because there should
most likely be a dedicated place to store these tasks, not a tasks
graph. But I don't want to add more changes to this PR.
(that is to say: used to be like that, not changing it _yet_)