a callback-based tasks graph implementation
Instead of polling tasks, use callbacks.

Each task is supposed to set the ._tasks_wait event when it
finishes. Then the graph knows it can continue.

The main loop implementation is in execute() still, and most of it
has to do with handling edge cases of cancellation, and of
wait-after-fail.

There's a lot of bookkeeping, but similar to what we've had before:
the graph itself is implemented in terms of ._tasks = {id: <Task>},
._dependencies = {<Task1>: {<Task2>, <Task3>}}, and ._dependents
which is the opposite of dependencies.

Keeping .add_task, remove_task, add_dependency, remove_dependency
interfaces the same.

Logic moved outside of the tasks graph: (see other commits)
  - task delay/.execute_after is now part of task itself
  - task execution depending on subgraph: tasks now have a dependency
    on their subgraphs, and that is the way we make sure tasks only
    run when their subgraph is executed, as opposed to subgraph-specific
    logic inside of the graph itself

Next commits will also make sure that the tasks do run the callbacks
when they finish.