WorkflowTask: use callbacks when done
When a task is done, it is not just setting the state which tells
the graph to move forward, it is running the callbacks; that is
done just by setting the result on the WorkflowTaskResult.

Every WorkflowTask must set the result. It's that simple.

Also, turns out we don't need multiple classes of a WorkflowTaskResult.
One is enough.

Also a lot of cleanup. This is the most cleanup-ish commit.