Still wake the graph every second
This is mostly in case I missed something - wake the graph once
a second so that it can iterate forward.

If there's nothing to do, it will just do nothing and go back to
sleeping on the event.
This is as opposed to the previous implementation, who would, on
every iteration, go through all tasks anyway. This time, we're not
as wasteful.