amqp_client: increase processing frequency
Might as well process events more often, which WAY reduces latency for
small/short tasks.
This makes the realistic cases quite a bit faster, because in real
blueprints, there's quite a lot of sending events and logs. If we're
waiting 0.2 seconds for every log to be confirmed, that adds up.

And if there's no events to process, the loop goes back to sleep
quickly anyway - this seems to have negligible effect on idle cpu
usage...