WorkflowTest decorator: don't copy if not required
If the user didn't pass any resources to copy explicitly, we can
just use the blueprint file as-is. This saves quite a lot of
time & resources on copying stuff around.

We're not passing any resources to copy ever, but I'm still going
to keep this whole mess of an implementation, because this is
allowed to be used by plugin tests