RD-208 Replace tasks polling with callbacks (#589)
* callbacks on execution cancel

a way to register functions to be called when an execution is
cancelled.
I'll use this in the next commits. Just getting it out of
the way early.

* replace tasks_iter with a .tasks property

i'm not gonna use the iterator-to-look-at-all-tasks method
anymore. Instead, this property will be implemented. Updating usage
first.

* a callback-based tasks graph implementation

Instead of polling tasks, use callbacks.

Each task is supposed to set the ._tasks_wait event when it
finishes. Then the graph knows it can continue.

The main loop implementation is in execute() still, and most of it
has to do with handling edge cases of cancellation, and of
wait-after-fail.

There's a lot of bookkeeping, but similar to what we've had before:
the graph itself is implemented in terms of ._tasks = {id: <Task>},
._dependencies = {<Task1>: {<Task2>, <Task3>}}, and ._dependents
which is the opposite of dependencies.

Keeping .add_task, remove_task, add_dependency, remove_dependency
interfaces the same.

Logic moved outside of the tasks graph: (see other commits)
  - task delay/.execute_after is now part of task itself
  - task execution depending on subgraph: tasks now have a dependency
    on their subgraphs, and that is the way we make sure tasks only
    run when their subgraph is executed, as opposed to subgraph-specific
    logic inside of the graph itself

Next commits will also make sure that the tasks do run the callbacks
when they finish.

* implement hash for WorkflowTask

we want to put tasks into sets and use them as dict keys, so
implement the necessary interface for that. We've been using id
as the dict key before, so continue doing that.

* WorkflowTask: use callbacks when done

When a task is done, it is not just setting the state which tells
the graph to move forward, it is running the callbacks; that is
done just by setting the result on the WorkflowTaskResult.

Every WorkflowTask must set the result. It's that simple.

Also, turns out we don't need multiple classes of a WorkflowTaskResult.
One is enough.

Also a lot of cleanup. This is the most cleanup-ish commit.

* move .execute_after handling into tasks themselves

instead of handling it in the graph, the delays will be handled here.
This means the graph can just call the task and don't worry about it.
The task itself will return an async result as normal, but it will
only be called later.

Implemented using threads, which is less than perfect, but there's
going to be at most a small number of those at once, because
tasks are only delayed when they're waiting for a retry.

* task .get: remove from graph

when we are NOT using tasks graphs, and we run a task, that task
is still added to a "tasks graph", even though we're not using one.

This is so that these tasks can be awaited for inside of dispatch.py,
_execute_workflow_function.

And so, when the task is done, remove it from that graph, so that
we don't need to await on it anymore.

This is unfortunate and probably stupid, because there should
most likely be a dedicated place to store these tasks, not a tasks
graph. But I don't want to add more changes to this PR.
(that is to say: used to be like that, not changing it _yet_)

* rewrite execute_operation tests

i just think they're way more readable like this.

Also drop the sleep because the idea was stupid, this implementation
doesn't care if it's 1 second difference or .001 seconds difference.

* Stop asserting on the capabilities warning.

It's been deprecated for years now. Time to drop it.
Dropping this just because it's giving me trouble with py2.6. Not worth
the time to really investigate I think.

* Stop using rabbitmq in the tests

It was just these mocks away.

* WorkflowTest decorator: don't copy if not required

If the user didn't pass any resources to copy explicitly, we can
just use the blueprint file as-is. This saves quite a lot of
time & resources on copying stuff around.

We're not passing any resources to copy ever, but I'm still going
to keep this whole mess of an implementation, because this is
allowed to be used by plugin tests

* Tests: use a local types.yaml

Instead of downloading types.yaml from the internet on every test,
copy a minimal types.yaml over here, and use that.

This types.yaml includes just enough stuff to make the tests pass.

I was considering naming things 100% differently, eg. just "Root"
instead of "cloudify.nodes.Root", but then I'd have to rewrite
every test blueprint as well.

Anyway, this MASSIVELY reduces test runtime, as you can imagine, by
skipping a HTTP request to the internet on every testcase.

* Increase the default threadpool size

The local tasks threadpool is mostly used for waiting on I/O
(REST requests), so there's no problem with having a higher default
amount of those threads.

* fixup test_tasks_graph

similar to real tasks, test tasks need to also set the async result.

Also, interface changes such as replacing .tasks_iter with .tasks

* Still wake the graph every second

This is mostly in case I missed something - wake the graph once
a second so that it can iterate forward.

If there's nothing to do, it will just do nothing and go back to
sleeping on the event.
This is as opposed to the previous implementation, who would, on
every iteration, go through all tasks anyway. This time, we're not
as wasteful.

* amqp_client: increase processing frequency

Might as well process events more often, which WAY reduces latency for
small/short tasks.
This makes the realistic cases quite a bit faster, because in real
blueprints, there's quite a lot of sending events and logs. If we're
waiting 0.2 seconds for every log to be confirmed, that adds up.

And if there's no events to process, the loop goes back to sleep
quickly anyway - this seems to have negligible effect on idle cpu
usage...

* send_node_event_error_handler

This is the handler used if ignore_failure is set: it used to create
a task and... do nothing with it.
No, tasks need to be called to actually do something. So let's call it.

* remove the useless isinstance

there's no way a task can return a HandlerResult directly - those
are only returned from Handlers indeed.

The only case was this test, and in this test, it was so that
the task is not retried. So use a more direct way to express that
instead.

* test_operation_retry: no interval between retries

it was just wasting time. 4 seconds saved!

* kill_request: iterate over callbacks

* add a comment