replace `__qualname__` with `__name__`
qualname isn't there in py2, and we're still testing this on py2,
even though we're not running this code (= the dispatcher) on py2.

However it doesn't really matter, because qualname is equal to name
for top-level names, and names that aren't top-level (eg. a method)
we can't handle anyway (utils.get_func doesn't get those)