RD-1048 script_runner: ignore log messages unrelated to the test
That test used to assert on the logs directly, but there can be
a "Waiting for process XXX to end" message, depending how long
did the process take to run.

That causes test_logger to fail randomly, especially on Jenkins.

Instead, just ignore messages that are unrelated to the test.