RD-1414 Cluster client: also retry /archive 404s (#662)
We retry fileserver downloads with other cluster nodes, because
filesystem replication is asynchronous.
However /blueprints/<id>/archive is also a fileserver download,
but it's gated behind restservice so that its 404s dont necessarily
have the headers we're looking for. And in principle I suppose
error messages don't need to have that header.

So, also look at the URL when deciding whether to retry. Can't think
of a better way.

(same for snapshots, but of course snapshots aren't as important,
because we're not really downloading them all that much; but
blueprints, we do, in the upload_blueprint workflow)