New flags for status reevaluation of plugins update
A new `reevaluate_active_statuses` flag added to the
`PluginsUpdateClient.update_plugins()` function.  It's purpuse is to map
the executions' statuses to plugin updates' statuses before performing
plugins update procedure.  `failed` and any `*cancel*` statuses of the
executions will be mapped to `failed` statuses of the plugin updates,
while `terminated` - to `successful`.