Plugins and deployment update improvements (#713)
* CYBL-1107 Automatically fix inputs' types (#675)

* Inputs type casting functionality

In some rare (?) cases there might exist deployments with invalid type
of inputs (e.g. '20' stored where input type should be integer).  This
patch adds functionality to attempt to automatically cast types of input
values to expected (a.k.a. inputs auto-correction).

* Update rest_client to pass add_correct_types flag

* CYBL-1107 Add --auto-correct-types to plugins update (#678)

The flag is going to be passed to the underlying deployments update
flow.

* CYBL-1118 Automatically correct valid booleans (#697)

Extend types auto-correction functionality to boolean type.  Map
strings: `"true"` to boolean `True` and `"false"` to `False`.

* CYBL-1114 Reevaluate plugins update statuses (#699)

* New flags for status reevaluation of plugins update

A new `reevaluate_active_statuses` flag added to the
`PluginsUpdateClient.update_plugins()` function.  It's purpuse is to map
the executions' statuses to plugin updates' statuses before performing
plugins update procedure.  `failed` and any `*cancel*` statuses of the
executions will be mapped to `failed` statuses of the plugin updates,
while `terminated` - to `successful`.

* Update reevaluate_active_statuses description

The `reevaluate_active_statuses` flag is also passed to the deployment
updates.

* CYBL-1114 Pass reevaluate_active_statuses in update_with_existing_blueprint (#702)